Lost and Found App for Charusat University
This project is android application that mimics a “LOST AND FOUND” application useful to our university. Overall the application is built for people who have either lost or found an item in the university and wish to report it. The application compares the values in the database to find the item’s rightful owner.